added missing static_cast

+++ Engine/src/Engine/Layers/ImGuiLayer.cpp
@@ -63,7 +63,7 @@ namespace Engine {
 	void ImGuiLayer::End() {
 		ImGuiIO &io      = ImGui::GetIO();
 		Application &app = Application::Get();
-		io.DisplaySize   = ImVec2(static_cast<float>(app.GetWindow().GetWidth()), app.GetWindow().GetHeight());
+		io.DisplaySize   = ImVec2(static_cast<float>(app.GetWindow().GetWidth()), static_cast<float>(app.GetWindow().GetHeight()));
 
 		ImGui::Render();
 		ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
